requirements.txt
numpy==1.18.3
pandas==1.3.5
Pillow==9.2.0
scikit_learn==1.1.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.17.0
